requirements.txt
fastapi
uvicorn[standard]
python-multipart
PyMuPDF
sentence-transformers
scikit-learn
requests
python-dotenv
gradio